proyecto_api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
psycopg[binary]==3.1.20
marshmallow==3.19.0